fix: filter reasoning-only models from general chat
